T02 Completed exercise 2 of threads project
Note: The ready_list, semaphore waiters list which were sorted by priority
are not sorted now. At the effective time, the element having maximum
effective priority is calculated. This is done because effective priority is
a dynamic quantity and it might be different at the time the element was inserted
as opposed to a later time instant.

Also the commit message of previous commit has a mistake, it says T03.5 but is
in fact T02.5